Moved guard clause for preventing early rendering
Moved guard clause for preventing early rendering

The guard clause preventing the rendering of the game UI if the game is not ready has been moved. Additionally, a new game over condition has been added to check if the game is over. If the game is over, background music is paused and different sounds are played based on the winner. The game UI is rendered if the game is ready.