fix: Comment out touchstart event listeners
The touchstart event listeners are commented out as they don't seem to work properly when both click and touchstart events are used. However, it would be better to use touchstart for actual mobile devices.

feat: Decrease minPlayers value temporarily

The minPlayers value is temporarily set to 2 instead of 4 for testing purposes. It will ultimately be set to 1.